// src/Application/ApiCalls/generated/models/StringResultDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AXECIEL.GMAO.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { InfoSupp } from './InfoSupp';
import {
    InfoSuppFromJSON,
    InfoSuppFromJSONTyped,
    InfoSuppToJSON,
    InfoSuppToJSONTyped,
} from './InfoSupp';

/**
 * 
 * @export
 * @interface StringResultDTO
 */
export interface StringResultDTO {
    /**
     * 
     * @type {string}
     * @memberof StringResultDTO
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof StringResultDTO
     */
    value?: string | null;
    /**
     * 
     * @type {string}
     * @memberof StringResultDTO
     */
    role?: string | null;
    /**
     * 
     * @type {string}
     * @memberof StringResultDTO
     */
    username?: string | null;
    /**
     * 
     * @type {string}
     * @memberof StringResultDTO
     */
    name?: string | null;
    /**
     * 
     * @type {InfoSupp}
     * @memberof StringResultDTO
     */
    infoSupp?: InfoSupp;
}

/**
 * Check if a given object implements the StringResultDTO interface.
 */
export function instanceOfStringResultDTO(value: object): value is StringResultDTO {
    return true;
}

export function StringResultDTOFromJSON(json: any): StringResultDTO {
    return StringResultDTOFromJSONTyped(json, false);
}

export function StringResultDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): StringResultDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'value': json['value'] == null ? undefined : json['value'],
        'role': json['role'] == null ? undefined : json['role'],
        'username': json['username'] == null ? undefined : json['username'],
        'name': json['name'] == null ? undefined : json['name'],
        'infoSupp': json['infoSupp'] == null ? undefined : InfoSuppFromJSON(json['infoSupp']),
    };
}

export function StringResultDTOToJSON(json: any): StringResultDTO {
    return StringResultDTOToJSONTyped(json, false);
}

export function StringResultDTOToJSONTyped(value?: StringResultDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'value': value['value'],
        'role': value['role'],
        'username': value['username'],
        'name': value['name'],
        'infoSupp': InfoSuppToJSON(value['infoSupp']),
    };
}