// src/Application/ApiCalls/generated/models/PlanningDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AXECIEL.GMAO.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PlanningDTOOutput } from './PlanningDTOOutput';
import {
    PlanningDTOOutputFromJSON,
    PlanningDTOOutputFromJSONTyped,
    PlanningDTOOutputToJSON,
    PlanningDTOOutputToJSONTyped,
} from './PlanningDTOOutput';

/**
 * 
 * @export
 * @interface PlanningDTO
 */
export interface PlanningDTO {
    /**
     * 
     * @type {string}
     * @memberof PlanningDTO
     */
    message?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlanningDTO
     */
    code?: string | null;
    /**
     * 
     * @type {PlanningDTOOutput}
     * @memberof PlanningDTO
     */
    result?: PlanningDTOOutput;
}

/**
 * Check if a given object implements the PlanningDTO interface.
 */
export function instanceOfPlanningDTO(value: object): value is PlanningDTO {
    return true;
}

export function PlanningDTOFromJSON(json: any): PlanningDTO {
    return PlanningDTOFromJSONTyped(json, false);
}

export function PlanningDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanningDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'message': json['message'] == null ? undefined : json['message'],
        'code': json['code'] == null ? undefined : json['code'],
        'result': json['result'] == null ? undefined : PlanningDTOOutputFromJSON(json['result']),
    };
}

export function PlanningDTOToJSON(json: any): PlanningDTO {
    return PlanningDTOToJSONTyped(json, false);
}

export function PlanningDTOToJSONTyped(value?: PlanningDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'code': value['code'],
        'result': PlanningDTOOutputToJSON(value['result']),
    };
}